Implement PlaylistService
Also add PlaylistRepository and TrackRepository definitions,
and refactor `Repository#remove(T)` to accept the ID (`K`)
instead of the entity (`T`).

Note that there is currently the possibility of a race condition.
It's still undecided how this will be tackled. It might be better
to have the persistence layer indicate if an instance of `T`
exists already or cannot be found, instead of having a
single method `save(T)`.